[onert] Split EventWriter.cc into 4 files
Split EventWriter.cc into 4 files since this file actually contains
4 classes and helper functions for each.
